requirements.txt
tensorflow>=2.8.0
keras>=2.8.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
pandas>=1.3.0
Pillow>=8.0.0
gradio>=3.0.0
